Debug RPM - 2

+++ cpp/main/src/inputs.cpp
@@ -53,9 +53,9 @@ float ReadRPM(motor_direction_t* dir)
     currTime = millis();
     if((currTime - prevTime) > 1000)
     {
-        float diff = EncoderDataDecoder((encoderValue - prevEncoderValue), &dir_var);
+        long diff = EncoderDataDecoder((encoderValue - prevEncoderValue), &dir_var);
         Serial.println(diff);
-        rpm = (diff / (float)1024) * 60;
+        rpm = ((float)diff / (float)1024) * 60;
         prevEncoderValue = encoderValue;
         prevTime = currTime;
     }
